+++ APIBlox.AspNetCore/InternalHelpers.cs
@@ -8,5 +8,8 @@ namespace APIBlox.AspNetCore
         public static Action<RequestErrorObject> AlterRequestErrorObjectAction { get; set; }
 
         public static Func<object, object> EnsureResponseCompliesWithAction { get; set; } = objectResultValue => new { Data = objectResultValue };
+
+        public static string ErrorResponseContentType { get; set; } = "application/problem+json";
+
     }
 }